split mod loading from callbacks.lua, moved most files into appropriate subfolders, renamed rooms folder into room for consistency, removed require mod when not used
