+++ scripts/stageapi/extra/waves.lua
