+++ scripts/stageapi/grid/customdoors.lua
