+++ scripts/stageapi/loadOrder.lua
@@ -1,27 +1,32 @@
 return {
     "scripts.stageapi.basic",
-    "scripts.stageapi.basiclibrary",
-    "scripts.stageapi.overlays",
-    "scripts.stageapi.rooms",
-    "scripts.stageapi.customgrids",
-    "scripts.stageapi.extrarooms",
-    "scripts.stageapi.customdoors",
-    "scripts.stageapi.gridgfx",
-    "scripts.stageapi.roomgfx",
-    "scripts.stageapi.customstage",
-    "scripts.stageapi.stageOverride",
-    "scripts.stageapi.bosses",
-    "scripts.stageapi.transitions",
-    "scripts.stageapi.altRockOverride",
-    "scripts.stageapi.callbacks",
-    "scripts.stageapi.commands",
-    "scripts.stageapi.saveSystem",
     "scripts.stageapi.library",
-    "scripts.stageapi.roomEditorFeatures",
-    "scripts.stageapi.waves",
-    "scripts.stageapi.customFloorGen",
+    "scripts.stageapi.stage.overlays",
+    "scripts.stageapi.room",
+    "scripts.stageapi.grid.customgrids",
+    "scripts.stageapi.stage.extrarooms",
+    "scripts.stageapi.grid.customdoors",
+    "scripts.stageapi.grid.gridgfx",
+    "scripts.stageapi.stage.roomgfx",
+    "scripts.stageapi.stage.customstage",
+    "scripts.stageapi.stage.stageOverride",
+    "scripts.stageapi.stage.bosses",
+    "scripts.stageapi.stage.transitions",
+    "scripts.stageapi.grid.altRockOverride",
+    "scripts.stageapi.core.callbacks",
+    "scripts.stageapi.core.modLoading",
+    "scripts.stageapi.core.commands",
+    "scripts.stageapi.core.saveSystem",
+    "scripts.stageapi.grid.roomEditorFeatures",
+    "scripts.stageapi.extra.waves",
+    "scripts.stageapi.stage.customFloorGen",
     "scripts.stageapi.basementRenovator",
     "scripts.stageapi.reimpl",
     "scripts.stageapi.modcompat",
     "scripts.stageapi.changelogs",
+
+    -- Intentionally left files that originally were separate do blocks as
+    -- explicit entries in this file instead of having a submodules file like room.lua,
+    -- library.lua and reimpl.lua, to keep the load order exactly the same and avoid
+    -- unforeseen bugs
 }
