+++ scripts/stageapi/room/persistentEnts.lua
@@ -1,5 +1,4 @@
 local shared = require("scripts.stageapi.shared")
-local mod = require("scripts.stageapi.mod")
 
 --[[ options
 {
