+++ scripts/stageapi/room/roomMetadata.lua
