+++ scripts/stageapi/stage/extrarooms.lua
