+++ scripts/stageapi/stage/roomgfx.lua
